fix: use isFetched for accurate reaction count
